fix redis
