+++ backend/Social_network/Social_network/settings/base.py
@@ -31,6 +31,9 @@ INSTALLED_APPS = [
     # whitenoise
     # 'whitenoise.runserver_nostatic',
 
+    # async lib
+    'adrf',
+
     # apps
     'account.apps.AccountConfig',
     'album.apps.AlbumConfig',
