+++ backend/Social_network/tasks/tasks.py
@@ -20,21 +20,26 @@ def upload_photos():
     time.sleep(20)
 
 
-# @shared_task
-# def task_db(job_params):
-#     logger.info(f"Пытаюсь изменить пост")
-#     post = Post.objects.get(pk=job_params['id'])
-#     post.text = 'i am win'
-#     post.save()
-#     time.sleep(1)
-
 @shared_task
-def task_db(url):
-    logger.info("Делаю запрос на внешний апи")
-    r = requests.get(url=url)
-    time.sleep(10)
-    return r.json()
+def task_db(job_params):
+    logger.info(f"Пытаюсь изменить пост")
+    post = Post.objects.get(pk=job_params['id'])
+    post.text = 'i am win'
+    post.save()
+    time.sleep(5)
 
+
+# не работает асинхронно, так как view ждет результата
+# @shared_task
+# def task_db(url):
+#     logger.info("Делаю запрос на внешний апи")
+#     r = requests.get(url=url)
+#     time.sleep(10)
+#     return r.json()
+
+# получить результат в view:
+# res = task_db.delay(url=url)
+# res = res.get()
 
 @shared_task
 def task_1(task_type=1):
@@ -50,3 +55,11 @@ def task_2(task_type=1):
 
     time.sleep(task_type * 3)
     return True
+
+
+@shared_task
+def test_task(seconds):
+    logger.info("Поймал!")
+
+    time.sleep(seconds)
+    return {'data': 'I am work'}
