+++ backend/Social_network/tasks/views.py
@@ -1,29 +1,127 @@
+import json
+import time
+
+import requests
 from celery.result import AsyncResult
-from django.http import JsonResponse
-from django.views.decorators.csrf import csrf_exempt
+from django.db import transaction
+from django.http import HttpResponse
+from rest_framework import generics, status, views
+from rest_framework.exceptions import APIException
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from tasks.tasks import task_db, test_task
 
-# from tasks.tasks import create_task
+from post.models import Post
+from post.serializers import PostListSerializer, PostCreateSerializer
 
+import logging
 
-# с отправкой на email
-@csrf_exempt
-def run_task(request):
-    if request.POST:
-        task_type = request.POST.get("type")
-        task = create_task.delay(int(task_type))
-        return JsonResponse({"task_id": task.id}, status=202)
-
-    else:
-        task = create_task.delay(int(1))
-        return JsonResponse({"task_id": task.id}, status=202)
+log = logging.getLogger(__name__)
 
 
-@csrf_exempt
-def get_status(request, task_id):
-    task_result = AsyncResult(task_id)
-    result = {
-        "task_id": task_id,
-        "task_status": task_result.status,
-        "task_result": task_result.result
-    }
-    return JsonResponse(result, status=200)
+class CreatePostTransactionView(generics.ListCreateAPIView):
+    serializer_class = PostListSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_serializer_class(self):
+        if self.request.method == 'GET':
+            return PostListSerializer
+        else:
+            return PostCreateSerializer
+
+    def get_queryset(self):
+        return Post.objects.filter(user=self.request.user).prefetch_related('photo')
+
+    def post(self, request, *args, **kwargs):
+        serializer = PostCreateSerializer(data=request.data, context={'user': self.request.user})
+        serializer.is_valid(raise_exception=True)
+
+        # если происходит ошибка внутри блока - откатываются все операции блока
+        try:
+            with transaction.atomic():
+                # save instance
+                instance = serializer.save()
+                instance.save()
+
+                # create task params
+                job_params = {"id": instance.id}
+
+                # submit task for background execution
+                task = transaction.on_commit(lambda: task_db.delay(job_params))
+
+        except Exception as e:
+            raise APIException(str(e))
+
+        return Response(status=status.HTTP_201_CREATED, data=f'Пост успешно создан,'
+                                                             f'res={task.status}')
+
+
+class CreatePostView(generics.ListCreateAPIView):
+    serializer_class = PostListSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_serializer_class(self):
+        if self.request.method == 'GET':
+            return PostListSerializer
+        else:
+            return PostCreateSerializer
+
+    def get_queryset(self):
+        return Post.objects.filter(user=self.request.user).prefetch_related('photo')
+
+    def post(self, request, *args, **kwargs):
+        serializer = PostCreateSerializer(data=request.data, context={'user': self.request.user})
+        serializer.is_valid(raise_exception=True)
+
+        # То же самое, что и сверху, но без блока транзакции
+        instance = serializer.save()
+
+        job_params = {"id": instance.id}
+
+        task = task_db.delay(job_params)
+
+        return Response(status=status.HTTP_201_CREATED, data=f'Пост успешно создан,'
+                                                             f'res={task.status}')
+
+
+class PostEditView(generics.RetrieveUpdateDestroyAPIView):
+    serializer_class = PostCreateSerializer
+
+    def get_queryset(self):
+        pass
+
+    def update(self, request, *args, **kwargs):
+        pass
+
+    def delete(self, request, *args, **kwargs):
+        Post.objects.all().delete()
+        return Response(status=status.HTTP_200_OK, data='Пост успешно удален')
+
+
+class DigitalAPiView(views.APIView):
+
+    # пытаемся оправить асинхронный запрос на внешний апи
+    # он даже без библиотеки не блокирует приложение
+    def get(self, request):
+        url = 'http://numbersapi.com/random/year?json'
+        res = requests.get(url=url)
+        time.sleep(5)
+
+        # r = requests.get('http://numbersapi.com/random/year?json')
+        # r = requests.get('https://www.boredapi.com/api/activity/')
+        # english
+        # r = requests.get('https://api.dictionaryapi.dev/api/v2/entries/en/fuck')
+        # food
+        # https: // spoonacular.com / food - api / pricing
+        # a lot of api, need token
+        # r = requests.get('https://api.api-ninjas.com/v1/cars?')
+
+        return Response(status=status.HTTP_200_OK, data=res.json())
+
+
+class RedisTestView(views.APIView):
+
+    def get(self, request):
+        res = test_task.delay(5)
+        return Response(status=status.HTTP_200_OK, data=res.json())
+
